+++ cluster.yaml
@@ -3,3 +3,6 @@ instance_type: t3.large
 min_instances: 3
 max_instances: 4
 telemetry: false
+
+# TODO nested?
+image_manager: 764403040460.dkr.ecr.us-west-2.amazonaws.com/cortexlabs/manager:latest
